merged in one version of testing.
I still need to merge in the latest version that has the correct
constriant function. Also in this commit, I made lapack and blas optional
with cmake and preprocessor flags @nzamb222. I need to implement a home
made version of psuedo inverse still. the one I put in there just copys
the array.